requirements.txt
google-generativeai
Pypdfium2
pillow
numpy
pandas
gradio